# swagger/swagger_api_docs.json
{
  "swagger": "2.0",
  "info": {
    "title": "ADAD Backend API",
    "description": "This document was generated from the branch main.\n\nIf you are a contributor to this project, you may access it here: https://paw.cloud/account/teams/5342920/projects/401341",
    "version": "v1.0.0",
    "contact": {
      "name": "Willian Santa Ana",
      "url": "https://paw.cloud/account/teams/5342920"
    }
  },
  "host": "localhost:3000",
  "schemes": [
    "http"
  ],
  "basePath": "/",
  "paths": {
...
    "/users": {
      "get": {
        "summary": "Get users with pagination",
        "description": "This endpoint retrieves a paginated list of users.\n\n### Query Parameters\n\n| Parameter | Type     | Description                           |\n|-----------|----------|---------------------------------------------|\n| `limit`   | `number` | The maximum number of users to return per page. Default: `10`. |\n| `page`    | `number` | The page number to retrieve. Default: `1`.  |",
        "operationId": "f418c216-09f3-46d0-afe7-7b535ad34d0e",
        "parameters": [
          {
            "type": "string",
            "default": "10",
            "name": "limit",
            "required": false,
            "in": "query"
          },
          {
            "type": "string",
            "default": "1",
            "name": "page",
            "required": false,
            "in": "query"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        }
      },
      "post": {
        "summary": "Create a New User",
        "description": "This endpoint creates one or more new users with the provided details.\n\n\n### Request Body\nThe body of the request must be in JSON format and include the following fields:\n\n| Field          | Type     | Description                          |\n|-----------------|----------|--------------------------------------|\n| `first_name`   | `string` | The first name of the user.          |\n| `last_name`    | `string` | The last name of the user.           |\n| `year_of_birth`| `number` | The year of birth of the user.       |\n| `job`          | `string` | The occupation of the user.          |\n| `reviews`      | `array`  | A list of reviews related to the user. |\n\n",
        "operationId": "d5f0721d-f286-4db2-8d00-365b6cff131c",
        "consumes": [
          "text/plain"
        ],
        "parameters": [
          {
            "required": false,
            "schema": {
              "type": "string",
              "default": "{\n  \"first_name\": \"Willian\",\n  \"last_name\": \"Santana\",\n  \"year_of_birth\": 1994,\n  \"job\": \"Software Engineer\",\n  \"reviews\": []\n}"
            },
            "in": "body",
            "name": "body"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        }
      }
    },
...
    "/books/comments": {
      "get": {
        "summary": "Retrieve Books with Comments",
        "description": "This endpoint retrieves a paginated list of books that have comments associated with them.\n\n### Query Parameters\n\n| Parameter | Type     | Description                           |\n|-----------|----------|---------------------------------------------|\n| `limit`   | `number` | The maximum number of books to return per page. Default: `10`. |\n| `page`    | `number` | The page number to retrieve. Default: `1`.  |",
        "operationId": "be77e60c-4b2c-4fda-a1cf-8157cab456ee",
        "parameters": [
          {
            "type": "string",
            "default": "10",
            "name": "limit",
            "required": false,
            "in": "query"
          },
          {
            "type": "string",
            "default": "1",
            "name": "page",
            "required": false,
            "in": "query"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        }
      }
    },
...
    "/comments": {
      "post": {
        "summary": "Create a New Comment",
        "description": "This endpoint creates a new comment for a specific book and user.\n\n## Request Body\nThe body of the request must be in JSON format and include the following fields:\n\n| Field      | Type     | Description                                     |\n|------------|----------|-------------------------------------------------|\n| `user_id`  | `string` | The unique identifier of the user making the comment. |\n| `book_id`  | `string` | The unique identifier of the book being commented on. |\n| `comment`  | `string` | The content of the comment.                     |",
        "operationId": "d9a29cf6-867f-47d2-a55a-a517bd509ed0",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "required": false,
            "schema": {
              "type": "string",
              "default": "{\"user_id\":\"6739b824e4c4f7929dbbf31c\",\"book_id\":\"6739b82ae4c4f7929dbbf31d\",\"comment\":\"Comment test\"}"
            },
            "in": "body",
            "name": "body"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        }
      }
    },
...
    "/bookstores/near": {
      "get": {
        "summary": "Retrieve Bookstores Near a Location",
        "description": "This endpoint retrieves a list of bookstores near a specific geographic location, based on provided coordinates.\n\n### Query Parameters\n\n| Parameter     | Type     | Description                                            |\n|---------------|----------|--------------------------------------------------------|\n| `latitude`    | `number` | The latitude of the location.                          |\n| `longitude`   | `number` | The longitude of the location.                         |\n| `maxDistance` | `number` | (Optional) The maximum distance (in meters) to search within. Default: `5000` (5 km). |\n\n&nbsp;\n\n### Query Parameters\n\n| Parameter | Type     | Description                           |\n|-----------|----------|---------------------------------------------|\n| `limit`   | `number` | The maximum number of books to return per page. Default: `10`. |\n| `page`    | `number` | The page number to retrieve. Default: `1`.  |",
        "operationId": "71ef348b-df19-4d33-a9a4-fd79fdd763aa",
        "parameters": [
          {
            "type": "string",
            "default": "38.7105419551935",
            "name": "latitude",
            "required": false,
            "in": "query"
          },
          {
            "type": "string",
            "default": "-9.14421890064127",
            "name": "longitude",
            "required": false,
            "in": "query"
          },
          {
            "type": "string",
            "default": "5000",
            "name": "maxDistance",
            "required": false,
            "in": "query"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        }
      }
    },
    "/bookstores/near-route": {
      "post": {
        "summary": "Retrieve Bookstores Near a Route",
        "description": "This endpoint retrieves a list of bookstores located near a specified route. The route is defined as an array of geographic points (latitude and longitude), and the search radius is defined by `maxDistance`.\n\n### Request Body\nThe body of the request must be in JSON format and include the following fields:\n\n| Field         | Type     | Description                                                              |\n|---------------|----------|--------------------------------------------------------------------------|\n| `route`       | `array`  | An array of geographic points defining the route. Each point must include `latitude` and `longitude`. |\n| `maxDistance` | `number` | (Optional) The maximum distance (in meters) to search from the route. Default: `5000` (5 km). |\n\n&nbsp;\n\n### Query Parameters\n\n| Parameter | Type     | Description                           |\n|-----------|----------|---------------------------------------------|\n| `limit`   | `number` | The maximum number of books to return per page. Default: `10`. |\n| `page`    | `number` | The page number to retrieve. Default: `1`.  |",
        "operationId": "2dd41044-e5b4-427d-91c6-acbc22a93bf2",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "required": false,
            "schema": {
              "type": "string",
              "default": "{\"route\":[[-9.14421890064127,38.7105419551935],[-9.14217296415889,38.7155597377788]],\"maxDistance\":5000}"
            },
            "in": "body",
            "name": "body"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        }
      }
    },
    "/bookstores/near-count": {
      "get": {
        "summary": "Retrieve the Count of Bookstores Near a Location",
        "description": "This endpoint retrieves the total number of bookstores located within a specified distance from a given geographic location.\n\n### Query Parameters\n\n| Parameter     | Type     | Description                                            |\n|---------------|----------|--------------------------------------------------------|\n| `latitude`    | `number` | The latitude of the location.                          |\n| `longitude`   | `number` | The longitude of the location.                         |\n| `maxDistance` | `number` | (Optional) The maximum distance (in meters) to search within. Default: `5000` (5 km). |\n",
        "operationId": "c127cdbc-261a-4f12-a20a-e66f680f455c",
        "parameters": [
          {
            "type": "string",
            "default": "38.7105419551935",
            "name": "latitude",
            "required": false,
            "in": "query"
          },
          {
            "type": "string",
            "default": "-9.14421890064127",
            "name": "longitude",
            "required": false,
            "in": "query"
          },
          {
            "type": "string",
            "default": "5000",
            "name": "maxDistance",
            "required": false,
            "in": "query"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        }
      }
    },
    "/bookstores/fair/inside": {
      "get": {
        "summary": "Check if a User is Inside the Book Fair",
        "description": "This endpoint checks if a user is located within the defined boundaries of the book fair based on their geographic coordinates.\n\n### Query Parameters\n\n| Parameter  | Type     | Description                                  |\n|------------|----------|----------------------------------------------|\n| `latitude` | `number` | The latitude of the user's current location. |\n| `longitude`| `number` | The longitude of the user's current location. |",
        "operationId": "13b0c9c2-a0b2-47e9-b87f-cfc471e578d8",
        "parameters": [
          {
            "type": "string",
            "default": "38.72749043040882",
            "name": "latitude",
            "required": false,
            "in": "query"
          },
          {
            "type": "string",
            "default": "-9.155644342145884",
            "name": "longitude",
            "required": false,
            "in": "query"
          }
        ],
        "responses": {
          "default": {
            "description": "no response description was provided for this operation"
          }
        }
      }
    }
  },
...
}